Campus Mölndal Website
A group school project I have worked on together with Wictor Sundström, Alexander Olsson and Joshua Klein Velderman during a period of 5 weeks.

My part encompasses, among others, the scss files, as I mainly focused on the visual aspects of the site. My goal was to be aesthetically consistent while keeping the code clean, and complying with the UX and UI guidelines.